Revert mandatory definition of country.
Remove the requirement to define the country of each author. As most
papers at EduComp are from Brazilian authors, that would only cost
one line without effective consequences. If the authors are not
Brazilian, or if they want to add that information, they still can
define the country as usual.